LPS-54457 Move EhcacheConfigurationUtil from com.liferay.portal.cache.ehcache to com.liferay.portal.dao.orm.hibernate.region.
