+++ portal-impl/src/com/liferay/portal/dao/orm/hibernate/region/LiferayEhcacheRegionFactory.java
@@ -15,7 +15,6 @@
 package com.liferay.portal.dao.orm.hibernate.region;
 
 import com.liferay.portal.cache.ehcache.CacheManagerUtil;
-import com.liferay.portal.cache.ehcache.EhcacheConfigurationUtil;
 import com.liferay.portal.cache.ehcache.ModifiableEhcacheWrapper;
 import com.liferay.portal.kernel.cache.CacheListener;
 import com.liferay.portal.kernel.cache.CacheListenerScope;
